backend/requirements.txt
fastapi
uvicorn
boto3
pydantic
python-multipart
httpx
google-genai
